Convert stdin instead of returning usage.
This change makes it so the beni will work on STDIN when no arguments
are passed.  Also some defaults are set if no parameters are passed
either.

---

I just stumbled onto this project this weekend while working on
GopherGala.  I was planning to work on something similar when I was
alerted to your repo.  My original idea was to have the application work
on the STDIN sort of like `cat`, `sed`, `less` etc. do.

Anyway, I went ahead and created a stub version of the app using your
exported beni.Highlight function just to have something to enter. But
I planned to go ahead and do PR to add this functionality to your app.
So here it is.